fix(gux-dropdown): better handle asynchronously loaded options
Dropdown options loaded asynchronously were failing to have event handlers attached because the
handlers were set on individual elements during the elment bootstrap lifecyle. This switches event
handling to a delegation pattern, where all of the events are handled in a parent element and their
origin is checked in order to see how they should be handled.

COMUI-255