feat(modal): add dynamic modal size
This adds a "dynamic" size option to the modal component allowing the size of the modal to be
determined primarily by the size of it's content. The modal will be constrained so that it doesn't
come within a fixed distance of the edge of the viewport, to prevent overflow, but otherwise will
expand to the size of it's content.

COMUI-290