fix(pagination): drop-up instead of drop-down
The pagination component is almost always placed at the bottom of the page, but the item count
selection uses a dropdown, which is awkward in that position. This is a bit of a kludgy fix to make
the component drop-up instead, until we land COMUI-242, which should give us a better general
mechanism for this behavior.

COMUI-320